update phony, add mobile versions of util bytes functions that don't try to store anything
