+++ src/util/bytes_mobile.go
@@ -1,0 +1,13 @@
+//+build mobile
+
+package util
+
+// On mobile, just return a nil slice.
+func GetBytes() []byte {
+	return nil
+}
+
+// On mobile, don't do anything.
+func PutBytes(bs []byte) {
+	return
+}
