+++ src/util/util.go
@@ -6,7 +6,6 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 )
 
@@ -23,19 +22,6 @@ func LockThread() {
 // A wrapper around runtime.UnlockOSThread() so it doesn't need to be imported elsewhere.
 func UnlockThread() {
 	runtime.UnlockOSThread()
-}
-
-// This is used to buffer recently used slices of bytes, to prevent allocations in the hot loops.
-var byteStore = sync.Pool{New: func() interface{} { return []byte(nil) }}
-
-// Gets an empty slice from the byte store.
-func GetBytes() []byte {
-	return byteStore.Get().([]byte)[:0]
-}
-
-// Puts a slice in the store.
-func PutBytes(bs []byte) {
-	byteStore.Put(bs)
 }
 
 // Gets a slice of the appropriate length, reusing existing slice capacity when possible
